feat: Refactor UI components to use forwardRef
Converts `Card` and `Button` components to `React.forwardRef` to allow for ref forwarding. This improves their composability and allows parent components to access and manage DOM elements of these UI components.

Also includes minor improvements to the `Toast` component for clearer messages and adds a loading indicator (`Loader2`) to `FinancialsPage`.